Provide stock YML files for configuring production
Since any deployment is likely to be slightly different offer templates
for people to use. Another change in this commit is the addition of the
dotenv gem so that important settings can actually be pulled from .env
rather than the YAML files. Ultimately everything could be refactored from
config/ and then just one resource needs to be managed