Updating workflow to handle DNG images
This patch should go most of the way to making DNG images process with
an error. Because ImageMagick / GraphicsMagick can only decode RAW files
if you try to do the resize operation first it fails miserably. Simply
switching the order works as long as you retain the file extension to
provide a hint that you are handling a RAW file