Rework codebase to check for local copies
Instead of using a temporary file service just redo the logic of
the FileContentDatastream since this was already being overridden.
As an added bonus it means better compatibility down the road. For
best performance find some how to pass the local path that can also
be used without being explicit